Calculando IR a ser pago em todos os ativos comprados nos meses anteriores. Ainda precisa salvar na base e levar em consideração compensação de prejuízo.

+++ stocks-core/Constants/IncomeTaxesConstants.cs
@@ -10,7 +10,7 @@
         /// <summary>
         /// Alíquota para o pagamento de imposto de renda sob ações.
         /// </summary>
-        public const int IncomeTaxesForStocks = 15;
+        public const decimal IncomeTaxesForStocks = 15;
 
         /// <summary>
         /// Alíquota para o pagamento de imposto de renda sob operações daytrade.
